requirements.txt
asgiref==3.5.2
autopep8==1.6.0
azure-core==1.26.4
azure-storage-blob==12.16.0
backports.zoneinfo==0.2.1
boto3==1.26.113
botocore==1.29.113
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dj-database-url==1.0.0
Django==4.0.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
jmespath==1.0.1
Pillow==9.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.30.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.2.0
